Separate lexer and parser sources
